apply refactoring to tracer
